+++ packages/analytics/src/Providers/AWSPinpointProvider.ts
@@ -14,7 +14,8 @@ import {
     ConsoleLogger as Logger, 
     ClientDevice, 
     Platform, 
-    Credentials
+    Credentials,
+    Signer
 } from '@aws-amplify/core';
 import * as MobileAnalytics from 'aws-sdk/clients/mobileanalytics';
 import * as Pinpoint from 'aws-sdk/clients/pinpoint';
@@ -42,6 +43,7 @@ export default class AWSPinpointProvider implements AnalyticsProvider {
     private mobileAnalytics;
     private pinpointClient;
     private _sessionId;
+    private _sessionStartTimestamp;
     private _buffer;
     private _clientInfo;
 
@@ -169,6 +171,9 @@ export default class AWSPinpointProvider implements AnalyticsProvider {
             return Promise.resolve(false);
         } 
         const timestamp = new Date().getTime();
+        // attach the session and eventId
+        this._generateSession(params);
+        params.event.eventId = uuid();
 
         Object.assign(params, { timestamp, config: this._config, credentials });
 
@@ -177,6 +182,44 @@ export default class AWSPinpointProvider implements AnalyticsProvider {
             return this._send(params);
         } else {
             return this._putToBuffer(params);
+        }
+    }
+
+    private _generateSession(params) {
+        this._sessionId = this._sessionId || uuid();
+        const { event } = params;
+
+        switch (event.name) {
+            case '_session_start':
+                // refresh the session id and session start time
+                this._sessionStartTimestamp = new Date().getTime();
+                this._sessionId = uuid();
+                event.session = {
+                    Id: this._sessionId,
+                    StartTimestamp: new Date(this._sessionStartTimestamp).toISOString()
+                };
+                break;
+            case '_session_stop':
+                const stopTimestamp = new Date().getTime();
+                this._sessionStartTimestamp = this._sessionStartTimestamp || new Date().getTime();
+                this._sessionId = this._sessionId || uuid();
+                event.session = {
+                    Id: this._sessionId,
+                    Duration: stopTimestamp - this._sessionStartTimestamp,
+                    StartTimestamp: new Date(this._sessionStartTimestamp).toISOString(),
+                    StopTimestamp: new Date(stopTimestamp).toISOString()
+                };
+                this._sessionId = undefined;
+                this._sessionStartTimestamp = undefined;
+                break;
+            default:
+                this._sessionStartTimestamp = this._sessionStartTimestamp || new Date().getTime();
+                this._sessionId = this._sessionId || uuid();
+                event.session = {
+                    Id: this._sessionId,
+                    StartTimestamp: new Date(this._sessionStartTimestamp).toISOString()
+                };
+                break;
         }
     }
 
@@ -199,43 +242,55 @@ export default class AWSPinpointProvider implements AnalyticsProvider {
         }
     }
 
-    /**
-     * @private
-     * @param params 
-     */
-    private async _startSession(params) {
-        // credentials updated
+    private _generateBatchItemContext(params) {
         const { event, timestamp, config, credentials } = params;
-        const { name, attributes, metrics } = event;
+        const { name, attributes, metrics, eventId, session } = event;
+        const { appId, endpointId } = config;
 
-        this._initClients(config, credentials);
+        const endpointContext = this._generateEndpointContext(config);
 
-        logger.debug('record session start');
-        this._sessionId = uuid();
-        const sessionId = this._sessionId;
-        
-        const clientContext = this._generateClientContext(config);
         const eventParams = {
-            clientContext,
-            events: [
-                {
-                    eventType: '_session.start',
-                    timestamp: new Date(timestamp).toISOString(),
-                    session: {
-                        id: sessionId,
-                        startTimestamp: new Date(timestamp).toISOString()
-                    },
-                    attributes,
-                    metrics
-                }
-            ]
+            ApplicationId: appId,
+            EventsRequest: {
+                BatchItem: {}
+            }
         };
-        
 
+        eventParams.EventsRequest.BatchItem[endpointId] = {};
+        const endpointObj = eventParams.EventsRequest.BatchItem[endpointId];
+        endpointObj['Endpoint'] = endpointContext;
+        endpointObj['Events'] = {};
+        endpointObj['Events'][eventId] = {
+            EventType: name,
+            Timestamp: new Date(timestamp).toISOString(),
+            Attributes: attributes,
+            Metrics: metrics,
+            Session: session
+        };
+
+        return eventParams;
+    }
+
+    private async _pinpointPutEvents(eventParams) {
+        logger.debug('pinpoint put events with params', eventParams);
         return new Promise<any>((res, rej) => {
-            this.mobileAnalytics.putEvents(eventParams, (err, data) => {
+            const request = this.pinpointClient.putEvents(eventParams);
+            // in order to keep backward compatiblity
+            // we are using a legacy api: /apps/{appid}/events/legacy
+            // so that users don't need to update their IAM Policy
+            // will use the formal one in the next break release
+            request.on('build', function() {
+                request.httpRequest.path = request.httpRequest.path + '/legacy';
+            });
+
+            request.send((err, data) => {
                 if (err) {
                     logger.debug('record event failed. ', err);
+                    logger.error(
+                        'Please ensure you have updated you Pinpoint IAM Policy' +
+                        'with the Action: \"mobiletargeting:PutEvents\" in order to' +
+                        'continue using AWS Pinpoint Service'
+                    );
                     res(false);
                 }
                 else {
@@ -250,44 +305,42 @@ export default class AWSPinpointProvider implements AnalyticsProvider {
      * @private
      * @param params 
      */
+    private async _startSession(params) {
+        // credentials updated
+        const { event, timestamp, config, credentials } = params;
+        this._initClients(config, credentials);
+
+        logger.debug('record session start');
+        const eventParams = this._generateBatchItemContext(params);
+        return this._pinpointPutEvents(eventParams);
+    }
+
+    /**
+     * @private
+     * @param params 
+     */
     private async _stopSession(params) {
         // credentials updated
         const { event, timestamp, config, credentials } = params;
-        const { name, attributes, metrics } = event;
-
         this._initClients(config, credentials);
 
         logger.debug('record session stop');
-    
-        const sessionId = this._sessionId ? this._sessionId : uuid();
-        const clientContext = this._generateClientContext(config);
-        const eventParams = {
-            clientContext,
-            events: [
-                {
-                    eventType: '_session.stop',
-                    timestamp: new Date(timestamp).toISOString(),
-                    session: {
-                        id: sessionId,
-                        stopTimestamp: new Date(timestamp).toISOString()
-                    },
-                    attributes,
-                    metrics
-                }
-            ]
-        };
-        return new Promise<any>((res, rej) => {
-            this.mobileAnalytics.putEvents(eventParams, (err, data) => {
-                if (err) {
-                    logger.debug('record event failed. ', err);
-                    res(false);
-                }
-                else {
-                    logger.debug('record event success. ', data);
-                    res(true);
-                }
-            });
-        });
+        const eventParams = this._generateBatchItemContext(params);
+        return this._pinpointPutEvents(eventParams);
+    }
+
+    /**
+     * @private
+     * @param params 
+     */
+    private async _recordCustomEvent(params) {
+        // credentials updated
+        const { event, timestamp, config, credentials } = params;
+        this._initClients(config, credentials);
+        
+        logger.debug('record event with params');
+        const eventParams = this._generateBatchItemContext(params);
+        return this._pinpointPutEvents(eventParams);
     }
 
     private async _updateEndpoint(params) : Promise<boolean> {
@@ -313,49 +366,6 @@ export default class AWSPinpointProvider implements AnalyticsProvider {
                     res(false);
                 } else {
                     logger.debug('updateEndpoint success', data);
-                    res(true);
-                }
-            });
-        });
-    }
-
-    /**
-     * @private
-     * @param params 
-     */
-    private async _recordCustomEvent(params) {
-        // credentials updated
-        const { event, timestamp, config, credentials } = params;
-        const { name, attributes, metrics } = event;
-
-        this._initClients(config, credentials);
-        
-        const clientContext = this._generateClientContext(config);
-        const sessionId = this._sessionId ? this._sessionId : undefined;
-        const eventParams = {
-            clientContext,
-            events: [
-                {
-                    eventType: name,
-                    timestamp: new Date(timestamp).toISOString(),
-                    session: {
-                        id: sessionId
-                    },
-                    attributes,
-                    metrics
-                }
-            ]
-        };
-
-        logger.debug('record event with params', eventParams);
-        return new Promise<any>((res, rej) => {
-            this.mobileAnalytics.putEvents(eventParams, (err, data) => {
-                if (err) {
-                    logger.debug('record event failed. ', err);
-                    res(false);
-                }
-                else {
-                    logger.debug('record event success. ', data);
                     res(true);
                 }
             });
@@ -441,34 +451,24 @@ export default class AWSPinpointProvider implements AnalyticsProvider {
      * @private
      * generate client context with endpoint Id and app Id provided
      */
-    private _generateClientContext(config) {
+    private _generateEndpointContext(config) {
         const { endpointId, appId } = config;
 
         const clientContext = config.clientContext || {};
         const clientInfo = this._clientInfo;
-        const clientCtx = {
-            client: {
-                client_id: clientContext.clientId || endpointId,
-                app_title: clientContext.appTitle,
-                app_version_name: clientContext.appVersionName,
-                app_version_code: clientContext.appVersionCode,
-                app_package_name: clientContext.appPackageName,
-            },
-            env: {
-                platform: clientContext.platform || clientInfo.platform,
-                platform_version: clientContext.platformVersion || clientInfo.version,
-                model: clientContext.model || clientInfo.model,
-                make: clientContext.make || clientInfo.make,
-                locale: clientContext.locale
-            },
-            services: {
-                mobile_analytics: {
-                    app_id: appId,
-                    sdk_name: Platform.userAgent
-                }
+
+        const endpointCtx = {
+            Demographic: {
+                Make: clientContext.make || clientInfo.make,
+                Model: clientContext.model || clientInfo.model,
+                Locale: clientContext.locale,
+                AppVersion: clientContext.appVersionName,
+                Platform: clientContext.platform || clientInfo.platform,
+                PlatformVersion: clientContext.platformVersion || clientInfo.version
             }
         };
-        return JSON.stringify(clientCtx);
+
+        return endpointCtx;
     }
 
     /**