+++ packages/aws-amplify-react-native/CHANGELOG.md
@@ -3,6 +3,100 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.7-unstable.1"></a>
+## [2.0.7-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.7-unstable.0...aws-amplify-react-native@2.0.7-unstable.1) (2018-10-08)
+
+
+### Bug Fixes
+
+* **aws-amplify-react:** jump to the initial state if not signed in ([d8779eb](https://github.com/aws/aws-amplify/commit/d8779eb))
+
+
+
+
+<a name="2.0.7-unstable.0"></a>
+## [2.0.7-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.6-unstable.0...aws-amplify-react-native@2.0.7-unstable.0) (2018-10-05)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="2.0.6"></a>
+## [2.0.6](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.6-unstable.0...aws-amplify-react-native@2.0.6) (2018-10-04)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="2.0.6-unstable.0"></a>
+## [2.0.6-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.5-unstable.0...aws-amplify-react-native@2.0.6-unstable.0) (2018-10-03)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="2.0.5"></a>
+## [2.0.5](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.5-unstable.0...aws-amplify-react-native@2.0.5) (2018-10-03)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="2.0.5-unstable.0"></a>
+## [2.0.5-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.4...aws-amplify-react-native@2.0.5-unstable.0) (2018-09-28)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="2.0.4"></a>
+## [2.0.4](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.4-unstable.0...aws-amplify-react-native@2.0.4) (2018-09-27)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="2.0.4-unstable.0"></a>
+## [2.0.4-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.3...aws-amplify-react-native@2.0.4-unstable.0) (2018-09-24)
+
+
+### Bug Fixes
+
+* **aws-amplify-react-native:** check contact before getting signed in ([5707577](https://github.com/aws/aws-amplify/commit/5707577))
+
+
+
+
+<a name="2.0.3"></a>
+## [2.0.3](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.3-unstable.0...aws-amplify-react-native@2.0.3) (2018-09-21)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="2.0.3-unstable.0"></a>
+## [2.0.3-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.1...aws-amplify-react-native@2.0.3-unstable.0) (2018-09-21)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="2.0.2"></a>
+## [2.0.2](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.1...aws-amplify-react-native@2.0.2) (2018-09-21)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
 <a name="2.0.1"></a>
 ## [2.0.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.1-unstable.3...aws-amplify-react-native@2.0.1) (2018-09-09)